Split out the mod file list from Mod and cleaned up a bunch of the warnings. Also added in the game framework for Lua.

+++ src/Basescape/BaseView.cpp
@@ -599,8 +599,8 @@ void BaseView::draw()
 					Surface *frame = _texture->getFrame((*craftIt)->getSkinSprite() + 33);		
 					int spriteWidthOffset= frame->getWidth()/2;  
 					int spriteHeightOffset= frame->getHeight()/2;	
-					int fx = (fac->getX() * GRID_SIZE) + ((fac->getRules()->getSizeX()) * GRID_SIZE) / 2.0 - spriteWidthOffset + p.x;
-					int fy = (fac->getY() * GRID_SIZE) + ((fac->getRules()->getSizeY()) * GRID_SIZE) / 2.0 + - spriteHeightOffset + p.y;	
+					int fx = static_cast<int>((fac->getX() * GRID_SIZE) + ((fac->getRules()->getSizeX()) * GRID_SIZE) / 2.0 - spriteWidthOffset + p.x);
+					int fy = static_cast<int>((fac->getY() * GRID_SIZE) + ((fac->getRules()->getSizeY()) * GRID_SIZE) / 2.0 + - spriteHeightOffset + p.y);
 					(*craftIt)->setBaseEscapePosition(Position(fx,fy,0));					
 					frame->blitNShade(this, fx, fy);
 					fac->addCraftForDrawing(*craftIt);
@@ -685,7 +685,11 @@ void BaseView::mouseOver(Action *action, State *state)
 	{
 		_selFacility = _facilities[_gridX][_gridY];
 		if ((_selFacility != 0)  && !(_selFacility->getCraftsForDrawing().empty())){
-			    Position mousePos(action->getRelativeXMouse()/action->getXScale(),action->getRelativeYMouse()/action->getYScale(),0);
+				Position mousePos(
+					static_cast<int>(action->getRelativeXMouse()/action->getXScale()),
+					static_cast<int>(action->getRelativeYMouse()/action->getYScale()),
+					0);
+
 				int dist=-1, newDist;
 				for (auto *craft : _selFacility->getCraftsForDrawing())
 				{
